server/requirements.txt
fastapi
uvicorn[standard]
pydub
pyttsx3
requests
faster-whisper
SpeechRecognition
python-dotenv
aiohttp